refactor(mlablocate*): use internal testing
This is not an external package and it's fine to just use internal testing.

It reduces the complexity a little bit.
